Use loadCSS to load css asynchronously in a way that is compatible with all browsers (mainly safari ios)
