+++ package.json
@@ -24,6 +24,7 @@
   },
   "homepage": "https://github.com/CaronaBoard/caronaboard#readme",
   "dependencies": {
+    "fg-loadcss": "^1.3.1",
     "firebase": "^3.5.3",
     "http-server": "^0.9.0"
   },
